--- a/app/controllers/guidances_controller.rb
+++ b/app/controllers/guidances_controller.rb
@@ -17,8 +17,7 @@ class GuidancesController < ApplicationController
     @guidances = Guidance.includes(:guidance_group, :themes)
                          .by_org(current_user.org).page(1)
     ensure_default_group(current_user.org)
-    @guidance_groups = GuidanceGroup.includes(:org)
-                                    .by_org(current_user.org).page(1)
+    @guidance_groups = GuidanceGroup.by_org(current_user.org).page(1)
   end
 
   # GET /org/admin/guidance/:id/admin_new