--- a/app/controllers/org_admin/phases_controller.rb
+++ b/app/controllers/org_admin/phases_controller.rb
@@ -57,7 +57,7 @@ module OrgAdmin
                  template: phase.template,
                  phase: phase,
                  prefix_section: phase.prefix_section,
-                 sections: phase.sections.order(:number)
+                 sections: phase.sections.includes(:template).order(:number)
                                          .select(:id, :title, :modifiable, :phase_id),
                  suffix_sections: phase.suffix_sections.order(:number),
                  current_section: Section.find_by(id: params[:section], phase_id: phase.id)