--- a/app/controllers/super_admin/orgs_controller.rb
+++ b/app/controllers/super_admin/orgs_controller.rb
@@ -11,7 +11,7 @@ module SuperAdmin
     def index
       authorize Org
       render 'index', locals: {
-        orgs: Org.with_template_and_user_counts.page(1)
+        orgs: Org.with_template_and_user_counts.page(1).includes(:contributors, :plans)
       }
     end