--- a/app/views/paginable/orgs/_index.html.erb
+++ b/app/views/paginable/orgs/_index.html.erb
@@ -31,7 +31,7 @@
               </button>
               <ul class="dropdown-menu" aria-labelledby="org-<%= org.id %>-actions">
                 <li class="nav-item"><%= link_to _('Edit'), admin_edit_org_path(org), class:'dropdown-item px-3' %></li>
-                <% unless org.has_users || org.template_count > 0 || org.has_contributors || org.has_plans %>
+                <% unless org.template_count > 0 || org.has_associations? %>
                   <li class="nav-item"><%= link_to _('Remove'), super_admin_org_path(org), data: {confirm: _("You are about to delete '%{org_name}'. Are you sure?") % { org_name: org.name}}, method: :delete, class:'dropdown-item px-3' %></li>
                 <% end %>
               </ul>