--- a/app/views/paginable/users/_index.html.erb
+++ b/app/views/paginable/users/_index.html.erb
@@ -93,8 +93,7 @@
                     <% end %>
                   </td>
                   <td class="text-center">
-                    <% presenter = IdentifierPresenter.new(identifiable: user) %>
-                    <% presenter.identifiers.each do |identifier| %>
+                    <% user.identifiers.each do |identifier| %>
                       <p><%= identifier.identifier_scheme.name %></p>
                     <% end %>
                   </td>