Changes proposed in this PR:
- This PR addresses some of the Bullet warnings that are triggered when using the app in development mode.
- Various ActiveRecord `.includes()` args have been added or removed to address these warnings.
- Some calls to `.length` have been changed to `.size` or `.any?` to address these warnings (which are sometimes false-positives).
  - `.length` brings the records into memory.
  - `.size` and `.any?` will only operate on the memory level if the records are already in memory. Otherwise, they will operate on the db level.
- `app/controllers/org_admin/templates_controller.rb` has been re-factored. In particular, the refactoring prevents N+1 queries related to the calls to `Template.published`.
- More details are documented within the individual commits of this PR.
- None of the actual returned results from these queries should be affected by this PR.

The following table compares the `development` branch to `aaron/bullet-fixes` by making requests to the various paths affected by this PR. The benchmarking was performed via ab - Apache HTTP server benchmarking tool alongside a recent (May 2024) db dump from the production environment of DMP Assistant. In both cases, 100 consecutive requests were performed to each path and the recorded result is the mean request time (ms).

Example request:
```
$ ab -n 100 -k -C "_dmp_roadmap_session=COOKIE_VALUE" -l http://127.0.0.1:3000/PATH/TO/PAGE
```


Path | mean request time (ms) before | mean request time (ms) now | % change
-- | -- | -- | --
/org_admin/plans (as org_admin type user) | 504.502 | 440.714 | -12.64%
/org_admin/plans (as super_admin type user) | 165.98 | 130.816 | -21.19%
/org_admin/templates | 328.156 | 162.383 | -50.52%
/org_admin/templates/organisational | 1022.935 | 969.703 | -5.20%
/org_admin/templates/customisable | 815.51 | 725.64 | -11.02%
/org_admin/templates/3925/phases/3994/edit | 83.591 | 85.359 | +2.12%
/org/admin/guidance/8/admin_index | 56.504 | 53.465 | -5.38%
/org/admin/users/admin_index | 144.602 | 139.837 | -3.30%
/paginable/guidance_groups/index/ALL | 14.816 | 12.931 | -12.72%
/public_plans | 135.203 | 99.571 | -26.35%
/super_admin/api_clients/new | 1625.653 | 653.083 | -59.83%
/super_admin/orgs | 1101.719 | 62.133 | -94.36%
/paginable/orgs/index/ALL | 3316.224 | 331.831 | -90%